### 2021-05-13 b0dc69b
Added a canny edge detector
