+++ particletracker/general/param_file_creator.py
@@ -26,9 +26,9 @@ def create_param_file(filename):
                               'C': [-23, -30, 30, 1],
                               'ad_mode': [True, ('True', 'False')]
                               },
-        'canny_edge':{'threshold1':100,
-                       'threshold2':200,
-                       'apertureSize':3
+        'canny_edge':{'threshold1':[100,0,255,1],
+                       'threshold2':[200,0,255,1],
+                       'apertureSize':[1,3,11,2]
                        },
         'colour_channel':{'colour':['red',('red','green','blue')]},
         'distance':{},
