### 2021-06-10 #3260 @Kevin-Mattheus-Moerman
@mikesmithlab  Some points to consider below.

- [ ] Can you please check that invalid DOI in the paper, see the reference check :point_up:
- [ ] Can you work on adding clear contribution guidelines (see [here](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) for more information)? I recommend adding a CONTRIBUTING.md file and to link to it from the README as well. In the contributing guidelines you can mention how people can contribute and also give guidelines on code style or variable naming if you like.

Recommendations:
- [ ] Since this package can create nice visual results to showcase performance/functionality I highly recommend including some visualizations in the JOSS paper. It would help readers to quickly grasp what the software is capable of.
- [ ] I also highly recommend adding something like a "graphical abstract" (an image showcasing the functionality) to the README. This way people that find your repository also quickly see what it is about.
- [ ] I recommend making the documentation link on your README an actual hyperlink, so use something like [https://particle-tracker.readthedocs.io/en/master/](https://particle-tracker.readthedocs.io/en/master/).  It is slightly more awkward at the moment to get to the link.
- [ ] I recommend actually linking to your license file in the README to, i.e. turn the text [the MIT licence](https://github.com/MikeSmithLabTeam/particletracker/blob/master/license.txt), like shown here.
- [ ] Once you have a CONTRIBUTING.md with contributing guidelines I recommend also adding a COC.md file which is the code of conduct for contributors. [This](https://www.contributor-covenant.org/version/2/0/code_of_conduct/code_of_conduct.md) may be a useful template to use from https://www.contributor-covenant.org/. I especially recommend this step if you aim for your project to grow and gain a lively contributor community.

If you update the paper you can call: `@whedon generate pdf` in a comment to automatically update the paper.
