### 2021-07-05 4d2d9fb
Update paper.bib
Corrected doi
