+++ paper.bib
@@ -77,7 +77,7 @@
     keywords = {bibtex-import},
     year = 2014,
     volume = 11,
-    doi = {https://doi.org/10.1038/nmeth.2808},
+    doi = {10.1038/nmeth.2808},
     url = {https://www.nature.com/articles/nmeth.2808},
 }
 
