+++ CONTRIBUTING.md
@@ -1,0 +1,18 @@
+## How to contribute to ParticleTracker
+
+#### **Did you find a bug?**
+
+* Bugs can be reported via the issue tracker (https://github.com/MikeSmithLabTeam/particletracker/issues)
+
+#### **Do you want to request a feature?**
+
+* If you would like to see a new feature please make a request in the issues tracker (https://github.com/MikeSmithLabTeam/particletracker/issues). If it seems that it would be of general usefulness we will happily consider implementing it.
+
+#### ** intend to add a new feature or change an existing one?**
+
+* If you'd like to extend / modify the functionality of ParticleTracker yourself please create a pull request. It is probably worth discussing this with us first since if the code is too specialist it might not be appropriate to include it. We may also be able to suggest a good way of going about implementing the feature. 
+
+#### **Do you have questions about the source code?**
+
+* We are happy to answer questions about the source code. Email: mikesmithlab@gmail.com or create an issue.
+
