+++ README.md
@@ -3,7 +3,7 @@
 ParticleTracker is a gui based particle tracking software that brings together a range of tools to make particle tracking simple. Full details on installation and use can be found in the documentation:
 
 ## Documentation 
-     https://particle-tracker.readthedocs.io/en/master/ 
+     Documentation (https://particle-tracker.readthedocs.io/en/master/) 
 
 ## Installation from github
     pip install git+https://github.com/MikeSmithLabTeam/particletracker
@@ -17,7 +17,7 @@ Add the following argument to setup.py setuptools.setup()
     dependency_links=['https://github.com/MikeSmithLabTeam/particletracker/tarball/repo/master#egg=package-1.0'],
     
 ## Contributions
-    Details about contributing to the ParticleTracker project can be found here: https://github.com/MikeSmithLabTeam/particletracker/blob/master/CONTRIBUTING.md
+    Details about contributing to the ParticleTracker project can be found here: (https://github.com/MikeSmithLabTeam/particletracker/blob/master/CONTRIBUTING.md)
 
 ## Licensing
 This project is licensed under the terms of the MIT licence (https://github.com/MikeSmithLabTeam/particletracker/blob/master/license.txt).
