### 2021-07-14 #3260 @mikesmithlab
COC now has contact method added.

Re: reviewers given the nature of the project - aimed at making things accessible to people not in the field, even those that can't code - anyone who can assess python code should be able to review.

If you are looking for more specialist opinions on particle tracking then a few ideas might be:

the authors of Trackpy https://github.com/soft-matter/trackpy this is a python particle tracking library which is actually used within our project to handle parts of the process. This is a mature library which allows the creation of particle tracking projects but does require significant coding.

There are a few authors who have written routines for particle tracking in other languages which are linked to here: http://www.physics.emory.edu/faculty/weeks//idl/ . There is one author who has written some form of GUI in IDL with a limited feature set but there doesn't appear to be a github repository.

Sorry I can't be of more help but will keep looking / thinking about alternatives
