### 2021-08-05 #3260 @Kevin-Mattheus-Moerman
@danielballan @caspervdw @alexlib @GenevieveBuckley would you be interested in reviewing this work for [the Journal of Open Source Software (JOSS)](https://joss.theoj.org/)? The [review process](https://joss.readthedocs.io/en/latest/submitting.html#the-review-process) takes place on GitHub and focuses on [the software](https://github.com/MikeSmithLabTeam/particletracker) and [the short paper](https://github.com/openjournals/joss-papers/blob/joss.03260/joss.03260/10.21105.joss.03260.pdf).

Note we can be very flexible regarding time needed to complete the review! If you can help please just let us know how much time you think we need.

If you are new to reviewing for JOSS, here is an example of a finished review process: https://github.com/openjournals/joss-reviews/issues/3323. As you can see reviewers have a checklist to complete which guides the review process.

Let me know if you have any questions.
