### 2021-08-06 #3260 @Kevin-Mattheus-Moerman
@alexlib starting in a month is fine if you are okay with that too. I have two other reviewers that could start at that time. We do [require](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) testing and encourage automated tests. The author @mikesmithlab might be able to comment on this.
