### 2021-08-14 MikeSmithLabTeam/particletracker#2 @alexlib
the `img sequence` option is not implemented
https://github.com/MikeSmithLabTeam/particletracker/blob/8b631e0d851146efa78885e114fa2baa0deb346f/particletracker/gui/main_gui.py#L103

the GUI opens with the options to choose only movies, `mp4`, `avi`, etc. no image extensions or instructions on how to choose an image sequence are given
https://github.com/openjournals/joss-reviews/issues/3611
