### 2021-08-14 MikeSmithLabTeam/particletracker#3 @alexlib
please create screencast videos
there are several steps which I could not reproduce following the documentation, but not sure if it's the lack of understanding of the procedures, or the order of actions or some combination of the installation and the misunderstanding of the steps. The screencasts of the examples would be useful.
Especially I have difficulty to visualize links for the colloids case, but also many other steps are not clear. https://github.com/openjournals/joss-reviews/issues/3611
