### 2021-08-20 #3611 @alexlib
I raised some issues about functionality on the respective repo
https://github.com/MikeSmithLabTeam/particletracker/issues/2
https://github.com/MikeSmithLabTeam/particletracker/issues/3

is this is the right way to handle this review?
