### 2021-08-25 MikeSmithLabTeam/particletracker#3 @mikesmithlab
This is a really nice idea. I will try and produce a vid for each of the examples
