### 2021-08-25 MikeSmithLabTeam/particletracker#2 @mikesmithlab
Not sure why this says Img Sequence. The image reading works through OpenCV and our intention was always to read from movies since this is what people record via a camera. Img Sequence might be a future addition but it is not something we intended to implement at this point. I will update the code to remove this confusion.
