### 2021-08-25 3365665
Update main_gui.py
Removed tool tips saying Img_Sequence since this is not implemented in this version
