+++ docs/source/rstfiles/overview.rst
@@ -43,26 +43,9 @@ Basic orientation for using the track_gui
 Read below or watch the video:
 
 .. raw:: html
-    <div id="ytplayer"></div>
-
-    <script>
-    // Load the IFrame Player API code asynchronously.
-    var tag = document.createElement('script');
-    tag.src = "https://www.youtube.com/player_api";
-    var firstScriptTag = document.getElementsByTagName('script')[0];
-    firstScriptTag.parentNode.insertBefore(tag, firstScriptTag);
-
-    // Replace the 'ytplayer' element with an <iframe> and
-    // YouTube player after the API code downloads.
-    var player;
-    function onYouTubePlayerAPIReady() {
-      player = new YT.Player('ytplayer', {
-      height: '360',
-      width: '640',
-      videoId: 'ajEp18opM-Y'
-      });
-    }
-    </script>
+    <div style="position: relative; padding-bottom: 56.25%; height: 0; overflow: hidden; max-width: 100%; height: auto;">
+        <iframe src="https://www.youtube.com/embed/ajEp18opM-Y" frameborder="0" allowfullscreen style="position: absolute; top: 0; left: 0; width: 100%; height: 100%;"></iframe>
+    </div>
     
     
 
