+++ docs/source/rstfiles/overview.rst
@@ -42,12 +42,12 @@ Basic orientation for using the track_gui
 -----------------------------------------
 Read below or watch the video:
 
-.. raw:: html
 
+.. raw:: html
     <div style="position: relative; padding-bottom: 56.25%; height: 0; overflow: hidden; max-width: 100%; height: auto;">
         <iframe src="https:~~//www.youtube.com/embed/ajEp18opM-Y" frameborder="0" allowfullscreen style="position: absolute; top: 0; left: 0; width: 100%; height: 100%;"></iframe>
     </div>
-    
+  
     
 
 The track_gui is divided into a few key areas:
