+++ docs/source/rstfiles/overview.rst
@@ -42,13 +42,7 @@ Basic orientation for using the track_gui
 -----------------------------------------
 Read below or watch the video:
 
-
-.. raw:: html
-
-    <div style="position: relative; padding-bottom: 56.25%; height: 0; overflow: hidden; max-width: 100%; height: auto;">
-        <iframe src="//www.youtube.com/embed/dQw4w9WgXcQ" frameborder="0" allowfullscreen style="position: absolute; top: 0; left: 0; width: 100%; height: 100%;"></iframe>
-    </div>
-  
+.. youtube:: dQw4w9WgXcQ  
     
 
 The track_gui is divided into a few key areas:
