+++ docs/source/rstfiles/overview.rst
@@ -43,9 +43,6 @@ Basic orientation for using the track_gui
 
 Read below or watch the video
 
-`Overview <https://www.youtube.com/watch?v=ajEp18opM-Y&ab_channel=MikeSmith>`_
-
-
 .. raw:: html
 
    <iframe width="560" height="315" src="https://www.youtube.com/embed/ajEp18opM-Y" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
