+++ docs/source/rstfiles/overview.rst
@@ -45,7 +45,7 @@ Read below or watch the video
 
 .. raw:: html
 
-   <iframe width="560" height="315" src="https://www.youtube.com/embed/ajEp18opM-Y" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
+   <iframe width="560" height="315" src="https://www.youtube.com/embed/ajEp18opM-Y&list=PL56zLBbX0yZZw18yyMM9tD0fLrobmdbJG&ab_channel=MikeSmith" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
     
 
 The track_gui is divided into a few key areas:
