### 2021-08-27 396ac6f
Update example2.rst
