### 2021-08-27 675ef57
Update example3.rst
