+++ docs/source/rstfiles/examples/example3.rst
@@ -3,6 +3,11 @@
 Example 3 - Swelling Hydrogels
 ==============================
 
+Read below or watch the video:
+
+.. raw:: html
+
+   <iframe width="560" height="315" src="https://www.youtube.com/embed/i0CKKcwwyGY&list=PL56zLBbX0yZZw18yyMM9tD0fLrobmdbJG&index=4&ab_channel=MikeSmith" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
 
 Open track_gui and load the movie hydrogel.mp4
 
