### 2021-08-27 81d3b40
Update example4.rst
