+++ docs/source/rstfiles/examples/example1.rst
@@ -13,7 +13,7 @@ Read below or watch the video
 
 .. raw:: html
 
-   <iframe width="560" height="315" src="https://www.youtube.com/embed/PjNNk0kU7EE&list=PL56zLBbX0yZZw18yyMM9tD0fLrobmdbJG&index=2&ab_channel=MikeSmith" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
+   <iframe width="560" height="315" src="https://www.youtube.com/embed/PjNNk0kU7EE" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
     
 
 
