### 2021-08-27 8ce718e
Update example2.rst
