### 2021-08-27 72fe8e0
Update example3.rst
