### 2021-08-27 a6a9415
Update example4.rst
