### 2021-08-31 503440c
Update example6.rst
