+++ docs/source/rstfiles/examples/example6.rst
@@ -2,6 +2,12 @@
 
 Example 6 - Working with the final data in a Jupyter Notebook
 =============================================================
+Read below or watch the video
+
+.. raw:: html
+
+   <iframe width="560" height="315" src="https://www.youtube.com/embed/BhGaWuiPADg" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
+
 
 If you used the precompiled executable of ParticleTracker you may not have 
 set up the details described in installation for working with a Jupyter Notebook.
