### 2021-09-09 540dc1c
Update example4.rst
reloaded example4 and updated link
