### 2021-09-09 MikeSmithLabTeam/particletracker#3 @mikesmithlab
> Example 4 is without audio. https://www.youtube.com/watch?v=iyeOTgB47yk&list=PL56zLBbX0yZZw18yyMM9tD0fLrobmdbJG&index=5

This has been resolved
