### 2021-09-09 #3611 @mikesmithlab
> @mikesmithlab are you able to add clearly documented automated (or manual) tests?

In our installation instructions on readthedocs:
https://particle-tracker.readthedocs.io/en/master/rstfiles/installation.html

the last section says:
"Verifying the installation
To verify that the installation is working correctly you should read the getting started section which explains how to launch the software for both the precompiled and python versions. Once you have done this there are 5 example videos which also act as tutorials which should enable you to verify and test the core functionality of the software."

The tests are not automated but in examples 2,3,4&5 one simply opens the gui, loads a settings file and then clicks the process button. Each example has its own page and walk through video which explains and shows what the expected behaviour / output should be.

Could you clarify what changes are being asked for as I'd assumed that this would satisfy the manual test request?
