### 2021-09-14 #3611 @nkeim
A couple quick comments related to my remaining checklist items for the paper:
1. The authors might want to double-check the author initials, at least for trackpy; "et al" might be appropriate. Also, I recommend citing the [trackpy Zenodo DOI](https://zenodo.org/record/3492186#.YUDl3C1h3UY) for the sake of automated citation-tracking.
2. There are sophisticated particle-tracking GUIs for FIJI (ImageJ) that seem to have wide adoption especially in the life sciences. [TrackMate](https://imagej.net/plugins/trackmate/) is one prominent example. I think it's worth citing at least one because it's another high-quality example of the state of the field, even though the overlap with ParticleTracker's intended user base may not be large.

I'm nearly done with my review (it's great!) but first need to create some issues on the software repository (and possibly mention a couple other things here).
