### 2021-09-15 MikeSmithLabTeam/particletracker#4 @nkeim
Mac supported?
Since the documentation is aimed at non-experts, it would be best to state explicitly that the software works on the Mac (I use a Mac and did not encounter any critical problems with ParticleTracker). There might be a weaker, but still helpful, way to say this if the Mac is not officially supported.

This issue stems from my JOSS review (https://github.com/openjournals/joss-reviews/issues/3611) but I do not consider it critical for publication in that journal.
