### 2021-09-15 MikeSmithLabTeam/particletracker#5 @nkeim
Link to current trackpy version
When installing ParticleTracker, pip installs the latest version of trackpy. This is probably good policy since major API changes in trackpy are quite rare. However the [overview page](https://particle-tracker.readthedocs.io/en/master/rstfiles/overview.html) (and maybe others) links only to the previous version (v0.4.2). It would make more sense to link to the [GitHub repo](https://github.com/soft-matter/trackpy) or to http://soft-matter.github.io/trackpy/stable

This issue stems from my JOSS review (https://github.com/openjournals/joss-reviews/issues/3611) but I do not consider it critical for publication in that journal.
