### 2021-09-15 MikeSmithLabTeam/particletracker#6 @nkeim
Orientation video not playing
The video at https://particle-tracker.readthedocs.io/en/master/rstfiles/overview.html does not play in Safari or Chrome. The text is still sufficient to learn the basic features of the software. Other videos in the documentation do seem to work.

This issue stems from my JOSS review (https://github.com/openjournals/joss-reviews/issues/3611) but I do not consider it critical for publication in that journal.
