### 2021-09-15 MikeSmithLabTeam/particletracker#7 @nkeim
How to use difference and rate postprocessing steps?
When looking at the example colloids results, I could not get the `difference` or `rate` postprocessing steps to work — I would just get a "difference error" or "rate error" message. For example, `column_name='x', output_name='vx', fps=50.0, span=3`. Is there a step I'm missing?

This issue stems from my JOSS review (https://github.com/openjournals/joss-reviews/issues/3611) but I do not consider it critical for publication in that journal.
