### 2021-09-15 MikeSmithLabTeam/particletracker#8 @nkeim
Glitch with `mask_rectangle`
When setting up a mask, I add a `mask_rectangle`, check the box to enable interactive drawing, and draw the region. If I then forgot to uncheck the box before adding another `mask_rectangle`, the first region I drew remains on the display and cannot be removed.

This issue stems from my JOSS review (https://github.com/openjournals/joss-reviews/issues/3611) but I do not consider it critical for publication in that journal.
