### 2021-09-15 MikeSmithLabTeam/particletracker#9 @nkeim
Preprocessed image doesn't work when "Use part processed" is turned on
Clicking "Captured image" to view the preprocessed image does nothing when "Use part processed" is turned on. This is confusing to a new user — perhaps an error message, or graying out the button, would help.

This issue stems from my JOSS review (https://github.com/openjournals/joss-reviews/issues/3611) but I do not consider it critical for publication in that journal.
