### 2021-09-15 MikeSmithLabTeam/particletracker#10 @nkeim
Crash when using `smooth` post-processing
Attempting to run the `smooth` post processing on the example eye-tracking data crashes the GUI (reproduced multiple times).

This issue stems from my JOSS review (https://github.com/openjournals/joss-reviews/issues/3611) but I do not consider it critical for publication in that journal.
