### 2021-09-15 MikeSmithLabTeam/particletracker#11 @nkeim
Crash when switching to a new movie
After completing the "Eyes" tutorial I moved on to "Colloids" by opening `colloids.mp4`. The GUI crashed with this traceback (perhaps because I had "Use part processed" still enabled?):

```
File /Users/nkeim/projects/joss-reviews/testdata/colloids.hdf5 does not exist
Error in DataStore.load
Postprocessing:   0%|                                                                                                 | 0/1 [00:00<?, ?it/s]
Error in general.dataframes
'DataStore' object has no attribute 'df'
Error in DataStore.save
Traceback (most recent call last):
  File "/Users/nkeim/opt/anaconda3/envs/particle/lib/python3.9/site-packages/particletracker/gui/main_gui.py", line 246, in <lambda>
    self.frame_selector.rangeChanged.connect(lambda x: self.param_change(x))
  File "/Users/nkeim/opt/anaconda3/envs/particle/lib/python3.9/site-packages/particletracker/gui/main_gui.py", line 340, in param_change
    self.update_viewer()
  File "/Users/nkeim/opt/anaconda3/envs/particle/lib/python3.9/site-packages/particletracker/gui/main_gui.py", line 391, in update_viewer
    annotated_img, proc_img = self.tracker.process_frame(frame_number, use_part=True)
  File "/Users/nkeim/opt/anaconda3/envs/particle/lib/python3.9/site-packages/particletracker/project/__init__.py", line 148, in process_frame
    self.pp.process(f_index=frame_num, use_part=use_part)
  File "/Users/nkeim/opt/anaconda3/envs/particle/lib/python3.9/site-packages/particletracker/postprocess/__init__.py", line 42, in process
    data.df = getattr(pm, method_name)(data.df, f_index=f,parameters=self.parameters, call_num=call_num)
AttributeError: 'DataStore' object has no attribute 'df'
zsh: abort      python
(particle) chou: /Users/nkeim/opt/anaconda3/envs/particle/lib/python3.9/multiprocessing/resource_tracker.py:216: UserWarning: resource_tracker: There appear to be 1 leaked semaphore objects to clean up at shutdown
  warnings.warn('resource_tracker: There appear to be %d '
```

This issue stems from my JOSS review (https://github.com/openjournals/joss-reviews/issues/3611) but I do not consider it critical for publication in that journal.
