### 2021-09-15 #3611 @nkeim
I've created the 8 above-referenced issues in the software repository, none of which I think should hold up publication. Aside from my remaining checklist items mentioned above, I am done and I am happy to recommend publication. I especially appreciate the tutorial videos and the range of representative examples, and I think the software is a good fit to the stated use case of basic particle tracking for research. Thanks!
