### 2021-09-16 MikeSmithLabTeam/particletracker#9 @mikesmithlab
Hi thanks for this.  This is by design. ie Preprocessed image shouldn't work when use part processed is turned on. The reason is that "use part proccesed" is no longer running the first 5 steps (including preprocessing) but is taking data from the already processed dataframe stored in videoname.hdf5 and then allowing you to interact with this via postprocessing and annotation. This is not only a way of speeding things up (ie you don't need to keep reprocessing things if they are working) but also necessary if you want to use a method (such as plotting a trajectory) that requires data from multiple frames. We disable the preprocess button so that people don't mistakenly think they are altering the preprocessing and wonder why nothing is happening. This is also indicated by the fact that the first 5 tabs should become greyed out when you select "use part processed". Hopefully that makes sense.
