### 2021-09-16 MikeSmithLabTeam/particletracker#11 @mikesmithlab
I'll look into this
