### 2021-09-16 MikeSmithLabTeam/particletracker#7 @mikesmithlab
Could you clarify if you have already run part_processed on the first 5 steps and then had selected the use_part_processed button. These features require data from multiple frames. They are only useable if you have a videoname.hdf5 file to work from?
