### 2021-09-16 d8e57b6
Update installation.rst
