### 2021-09-16 MikeSmithLabTeam/particletracker#4 @mikesmithlab
This is helpful thanks. I guess we are just saying we haven't tested it on Mac! Good to know it still works! I have added a clear statement about the operating systems. Saying it should work on Windows, Linux and Mac but that it has only been thoroughly tested on Windows 10 and Ubuntu Linux.
