### 2021-09-16 MikeSmithLabTeam/particletracker#11 @mikesmithlab
Looking at the error message and reading your comment the issue is this. Use_part_processed assumes you have already run the limited processing to create a videoname.hdf5 It then uses this datafile to run the postprocessing. So when you did the Eyes tutorial you created a eyes.hdf5 file. When you then opened colloids.mp4 the use_part_processed button means that the software immediately tried to access colloids.hdf5 to run a postprocessing command. The file doesn't exist though. Hence it crashes the software. I could perhaps make sure that when you open a movie it resets this use_part_processed toggle button which would then avoid this.
