### 2021-09-16 MikeSmithLabTeam/particletracker#10 @mikesmithlab
Do you have any more details or perhaps the output when it crashes as this is a bit difficult to troubleshoot?
