+++ paper.bib
@@ -1,10 +1,20 @@
+@misc{TrackMate,
+  author = {J. Tineveza, N. Perry, J. Schindelin, G.M. Hoopes, G.D. Reynolds, E. Laplantine, S.Y. Bednarek, S.L. Short, K.W. Eliceiri},
+  title = {TrackMate: An open and extensible platform for single-particle tracking},
+  year = {2016},
+  publisher = {Elsevier},
+  journal = {Methods},
+  url = {https://doi.org/10.1016/j.ymeth.2016.09.016}
+}
+
+
 @misc{Trackpy,
   author = {D. Allan, C.A. der Wel, N. Keim, T.A. Caswell et al.},
   title = {Trackpy: Fast, Flexible Particle-Tracking Toolkit},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://https://github.com/soft-matter/trackpy/tree/v0.4.2}
+  url = {https://zenodo.org/record/3492186#.YUMkp7hKjIV}
 }
 
 @article{opencv_library,
