+++ paper.bib
@@ -9,7 +9,7 @@
 
 
 @misc{Trackpy,
-  author = {D. Allan, C.A. der Wel, N. Keim, T.A. Caswell et al.},
+  author = {D.B. Allan, T. Caswell, N.C. Keim, C.M. van der Wel},
   title = {Trackpy: Fast, Flexible Particle-Tracking Toolkit},
   year = {2020},
   publisher = {GitHub},
