### 2021-09-16 #3611 @mikesmithlab
> A couple quick comments related to my remaining checklist items for the paper:
>
> 1. The authors might want to double-check the author initials, at least for trackpy; "et al" might be appropriate. Also, I recommend citing the [trackpy Zenodo DOI](https://zenodo.org/record/3492186#.YUDl3C1h3UY) for the sake of automated citation-tracking.
> 2. There are sophisticated particle-tracking GUIs for FIJI (ImageJ) that seem to have wide adoption especially in the life sciences. [TrackMate](https://imagej.net/plugins/trackmate/) is one prominent example. I think it's worth citing at least one because it's another high-quality example of the state of the field, even though the overlap with ParticleTracker's intended user base may not be large.
>
> I'm nearly done with my review (it's great!) but first need to create some issues on the software repository (and possibly mention a couple other things here).

I've checked initials, and changed / added the suggested citation to the paper.bib file.  I've resolved some of the issues above. The issue about smoothing needs some more info to resolve. The other two may be small bugs that I think I can resolve fairly easily. Thanks for your efforts, much appreciated.
