### 2021-09-16 MikeSmithLabTeam/particletracker#11 @nkeim
That makes sense. Yes, I agree that it's worthwhile to check and at least display an error dialog.
