### 2021-09-16 #3611 @nkeim
@mikesmithlab Thanks! One small final request: it's probably best practice to link to https://doi.org/10.5281/zenodo.3492186 instead of the zenodo.org page for trackpy. In any case I'm happy to sign off and I have checked all my boxes.
