### 2021-09-16 MikeSmithLabTeam/particletracker#7 @nkeim
Yes, this is with tracking completed and use_part_processed turned on.
