### 2021-09-16 8b2205d
use_part_button now unchecks when loading new video to avoid possible crash
