+++ particletracker/annotate/__init__.py
@@ -16,9 +16,9 @@ class TrackingAnnotator:
         self.data_filename=data_filename
         self.output_filename = self.cap.filename[:-4] + '_annotate.mp4'
 
-    def annotate(self, f_index=None, use_part=False, framerate=30):
-        frame = self.cap.read_frame(n=self.cap.frame_num)
+    def annotate(self, f_index=None, use_part=False):     
         if f_index is None:
+            frame = self.cap.read_frame(n=0)
             self.out = WriteVideo(filename=self.output_filename, frame=frame)
             data_filename = self.data_filename
         elif use_part:
