+++ particletracker/__init__.py
@@ -75,3 +75,5 @@ def track_batchprocess(moviefilter, settings,
         tracker.postprocess_select = postprocess
         tracker.annotate_select = annotate
         tracker.process(excel=excel)
+
+
