+++ paper.bib
@@ -14,7 +14,7 @@
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://zenodo.org/record/3492186#.YUMkp7hKjIV}
+  url = {https://doi.org/10.5281/zenodo.3492186}
 }
 
 @article{opencv_library,
