### 2021-09-16 #3611 @mikesmithlab
Not sure what I'm doing wrong. On the paper.bib file the url is "https://doi.org/10.5281/zenodo.3492186" but in the proof it reverts to https://zenodo.org........
