### 2021-09-17 90f6e74
Added extra arguments to the trackpy tracking method
