### 2021-09-17 714ece0
Resolved crop return bug
