+++ particletracker/gui/custom_combo_box.py
@@ -41,16 +41,16 @@ class ComboBoxAndButton(QWidget):
         draggable_list = self.parent.list_draggable_lists[tab_index]
         count = 1
 
-        if ok_to_duplicate_method_check(method):
-            if dynamic:
-                while method in draggable_list.method_list:
-                    method = method.split('*')[0] + '*' + str(count)
-                    count = count + 1
-                draggable_list.add_item(method, update=True)
-            else:
-                draggable_list.takeItem(0)
-                draggable_list.add_item(method)
-                draggable_list.send_signal()
+        #if ok_to_duplicate_method_check(method):
+        if dynamic:
+            while method in draggable_list.method_list:
+                method = method.split('*')[0] + '*' + str(count)
+                count = count + 1
+            draggable_list.add_item(method, update=True)
+        else:
+            draggable_list.takeItem(0)
+            draggable_list.add_item(method)
+            draggable_list.send_signal()
 
             
             
