### 2021-09-18 #3611 @Kevin-Mattheus-Moerman
@mikesmithlab is that citation issue resolved now? Let me know if I can help. Also can you check this potentially missing DOI :point_up: ?
