### 2021-09-18 #3611 @mikesmithlab
Citation issue seems to be resolved now
