### 2021-09-18 MikeSmithLabTeam/particletracker#8 @mikesmithlab
Have resolved this in the CurrentDevelopment branch. Once I have checked and resolved other issues you raised will merge with Main.
