+++ README.md
@@ -1,5 +1,3 @@
-#Quick check
-
 # ParticleTracker
 
 ParticleTracker is a gui based particle tracking software that brings together a range of tools to make particle tracking simple. Full details on installation and use can be found in the documentation:
