### 2021-09-21 b12ba3e
initial work on progress bar
