### 2021-09-21 cedd635
Status bar now tells user how to track processing
