### 2021-09-21 MikeSmithLabTeam/particletracker#10 @mikesmithlab
Smooth appears to be a legacy name which uses the rolling_mean. The example tutorial references using mean and I think this is preferred. Far better to have users consciously aware of what they are doing to their data than a magic button that smooths. I am going to remove this in the CurrentDevelopment version.
