+++ particletracker/postprocess/postprocessing_methods.py
@@ -1,3 +1,4 @@
+from PyQt5.QtCore import center
 import numpy as np
 import scipy.spatial as sp
 import trackpy as tp
@@ -627,255 +628,6 @@ def _get_class_subset(df, f, parameters, method=None):
     return subset_df
 
 
-'''
----------------------------------------------------------------------------------------------
-All these methods depend on information from other frames. ie they won't work unless
-multiple frames have been processed and you are using part.
----------------------------------------------------------------------------------------------
-'''
-def difference(df, f_index=None, parameters=None, call_num=None):
-    '''
-    Difference of a particles values on user selected column. 
-
-    Notes
-    -----
-    Returns the difference of a particle's values on a particular column at span separation in frames to a new column.
-    
-
-    column_name
-        Input column name
-    output_name
-        Column name for median data
-    span
-        number of frames over which to calculate rolling median
-    
-    
-    Args
-    ----
-
-    df
-        The dataframe in which all data is stored
-    f_index
-        Integer specifying the frame for which calculations need to be made.
-    parameters
-        Nested dictionary like object (same as .param files or output from general.param_file_creator.py)
-    call_num
-        Usually None but if multiple calls are made modifies method name with get_method_key
-
-    Returns
-    -------
-        updated dataframe including new column
-
-    '''
-    try:    
-        params = parameters['postprocess']
-        method_key = get_method_key('difference', call_num)
-        column = params[method_key]['column_name']
-        output_name = params[method_key]['output_name']
-        span = get_param_val(params[method_key]['span'])
-    
-        if output_name not in df.columns:
-            df[output_name] = np.nan
-
-        start=f_index-span - 1
-        if start < 0:
-            start = 0
-        
-        df_frames = df.loc[start:f_index,[column,'particle']]
-        df_diff=df_frames.groupby('particle')[column].diff(periods=span).transform(lambda x:x).to_frame(name=output_name)
-        #df_output.reset_index('particle', inplace=True)
-        df.loc[f_index,[output_name]]=df_diff.loc[f_index]
-    
-        return df
-    except Exception as e:
-        raise DifferenceError(e)
-
-def mean(df, f_index=None, parameters=None, call_num=None):
-    '''
-    Rolling mean of a particles values. 
-
-    Notes
-    -----
-    Returns the rolling mean of a particle's values to a new column. Useful
-    to reduce fluctuations or tracking inaccuracies.
-    
-
-    column_name
-        Input column name
-    output_name
-        Column name for mean data
-    span
-        number of frames over which to calculate rolling mean
-    
-    
-    Args
-    ----
-
-    df
-        The dataframe in which all data is stored
-    f_index
-        Integer specifying the frame for which calculations need to be made.
-    parameters
-        Nested dictionary like object (same as .param files or output from general.param_file_creator.py)
-    call_num
-        Usually None but if multiple calls are made modifies method name with get_method_key
-
-    Returns
-    -------
-        updated dataframe including new column
-
-
-    '''
-    
-
-    try:
-        params = parameters['postprocess']
-        method_key = get_method_key('mean', call_num)
-        column = params[method_key]['column_name']
-        output_name = params[method_key]['output_name']
-        span = get_param_val(params[method_key]['span'])
-
-        if output_name not in df.columns:
-            df[output_name] = np.nan
-
-        start=f_index-span - 1
-        if start < 0:
-            start = 0
-        
-        df_frames = df.loc[start:f_index,[column,'particle']]
-        df_output=df_frames.groupby('particle')[column].rolling(span).mean().transform(lambda x:x).to_frame(name=output_name)
-        df_output.reset_index('particle', inplace=True)
-        df.loc[f_index,[output_name]]=df_output.loc[f_index]
[…]
