### 2021-09-21 ba31ede
Improved doc strings for gui sections
