+++ particletracker/gui/main_gui.py
@@ -64,13 +64,17 @@ class MainWindow(QMainWindow):
                     self.settings_filename = self.movie_filename[:-4] + '_expt.param'
                 except:
                     print('tried to load current params - falling back on initial settings file')
+        
         self.open_tracker()
         self.setWindowTitle("Particle Tracker")
+
         self.main_panel = QWidget()
         self.main_layout = QHBoxLayout()  # Contains view and settings layout
         self.view_layout = QVBoxLayout()  # Contains image, image toggle button and frame slider
         self.settings_layout = QVBoxLayout()  # Contains tab widget with all tracking controls
-        self.init_ui(self.view_layout, self.settings_layout)
+
+        self.init_ui()#self.view_layout, self.settings_layout)
+        
         self.main_layout.addLayout(self.view_layout,3)
         self.main_layout.addLayout(self.settings_layout,2)
         self.main_panel.setLayout(self.main_layout)
@@ -79,7 +83,7 @@ class MainWindow(QMainWindow):
         self.showMaximized()
         
 
-    def init_ui(self, view_layout, settings_layout, reboot=True):
+    def init_ui(self):#, view_layout, settings_layout, reboot=True):
         if not hasattr(self, 'file_menu'):
             self.setup_menus_toolbar()
         else:
@@ -87,12 +91,20 @@ class MainWindow(QMainWindow):
             #and open a new video which has no .hdf5 file.
             self.use_part_button.setChecked(False)
 
-        self.setup_viewer(view_layout)# Contains image window, frame slider and spinbox.
-        self.setup_settings_panel(settings_layout)# Contains all widgets on rhs.
+        self.setup_viewer(self.view_layout)# Contains image window, frame slider and spinbox.
+        self.setup_settings_panel(self.settings_layout)# Contains all widgets on rhs of gui.
       
 
+        """------------------------------------------------------------------------------
+        ------------------------------------------------------------------------------
+        SETUP MENUS AND TOOLBAR
+        
+        Setup of all the menus and toolbars at the bottom. The statusbar at the bottom
+        is also initialised here.
+        ---------------------------------------------------------------------------------
+        ---------------------------------------------------------------------------------"""
+
     def setup_menus_toolbar(self):
-        #is for python gui, 
         dir,_ =os.path.split(os.path.abspath(__file__))
         resources_dir = os.path.join(dir,'icons','icons')
         #Use these lines when using pyinstaller.
@@ -196,7 +208,6 @@ class MainWindow(QMainWindow):
         self.process_menu.addAction(self.use_part_button)
         self.process_menu.addAction(process_button)
         
-
         docs = QAction('help', self)
         docs.triggered.connect(self.open_docs)
         about = QAction('about', self)
@@ -206,7 +217,6 @@ class MainWindow(QMainWindow):
         self.help_menu.addAction(about)
 
        
-
     def open_docs(self):
         webbrowser.open('https://particle-tracker.readthedocs.io/en/latest/', new=1, autoraise=True)
     
@@ -221,6 +231,17 @@ class MainWindow(QMainWindow):
             We are happy to consider feature requests."""
             )
         msg.show()
+
+
+
+        """-----------------------------------------------------------------------------------------------
+        --------------------------------------------------------------------------------------------------
+        SETUP VIEWER
+        
+        Setup for the viewer panel - LHS of Gui includes viewer, frame selector and capture v preprocessing
+        image toggle button.
+        --------------------------------------------------------------------------------------------------
+        ----------------------------------------------------------------------------------------------"""
 
     def setup_viewer(self, view_layout):
         '''
@@ -268,15 +289,19 @@ class MainWindow(QMainWindow):
         self.update_viewer()
         
 
-    '''
+    '''---------------------------------------------------------------
[…]
