### 2021-09-21 38294df
Improving docstrings and error messages
