### 2021-09-22 27db9a4
Added a very basic test script
