+++ tests/test_gui.py
@@ -1,0 +1,3 @@
+from particletracker import track_gui
+
+track_gui()
