### 2021-09-23 1e9a3e7
exposing additional options trackpy
