+++ docs/source/rstfiles/reference/launch_tracking.rst
@@ -1,0 +1,7 @@
+Launching Tracking Methods
+==========================
+
+.. automodule:: particletracker.__init__
+    :members:
+    :undoc-members:
+    :exclude-members: QListWidget, pyqtSignal
