### 2021-09-23 2b49b7a
Added a button to view the current frame of the dataframe in a popup widget
