### 2021-09-23 42a731a
pandas view has a button to save to csv
