+++ particletracker/customexceptions/postprocessor_error.py
@@ -102,6 +102,12 @@ class VoronoiError(PostprocessorError):
         self.error_msg = 'voronoi error'
         self.e = e
 
+class HexaticOrderError(PostprocessorError):
+    def __init__(self, e):
+        super().__init__(e)
+        self.error_msg = 'Hexatic error'
+        self.e = e
+
 class RemoveMaskedError(PostprocessorError):
     def __init__(self, e):
         super().__init__(e)
