### 2021-09-23 86be19c
added missing raise causing errors to not propogate
