### 2021-09-23 05eb6df
Added cmap_name option to annotation steps
