+++ particletracker/general/param_file_creator.py
@@ -166,6 +166,7 @@ def create_param_file(filename):
                 'cmap_column':'x',  #None
                 'cmap_max':[100.0,0,1000.0,0.1],#For dynamic
                 'cmap_min':[0.0,0.0,1000.0,0.1],
+                'cmap_name': 'jet',
                 'colour': (0, 255, 0),  # For static
                 'classifier_column': None,  # For static or dynamic
                 'classifier': [True, ('True','False')],
@@ -180,6 +181,7 @@ def create_param_file(filename):
                     'cmap_column':'x',#for dynamic
                     'cmap_max':[100.0,0,1000.0,0.1],#For dynamic
                     'cmap_min':[0.0,0.0,2000.0,0.1],
+                    'cmap_name': 'jet',
                     'colour': (0,255,0),#For static
                     'classifier_column': None,#For static or dynamic
                     'classifier': [True, ('True','False')],#For static or dynamic
@@ -189,6 +191,7 @@ def create_param_file(filename):
                     'cmap_column':'x',#For dynamic
                     'cmap_max':[100.0,0,1000.0,0.1],#For dynamic
                     'cmap_min':[0.0,0.0,1000.0,0.1],
+                'cmap_name': 'jet',
                     'colour': (0,255,0),#For static
                     'classifier_column': None,#For static or dynamic
                     'classifier': [True, ('True','False')],
@@ -201,6 +204,7 @@ def create_param_file(filename):
                     'cmap_column':'x',#For dynamic
                     'cmap_max':[100.0,0,1000.0,0.1],#For dynamic
                     'cmap_min':[0.0,0.0,1000.0,0.1],
+                'cmap_name': 'jet',
                     'colour': (64,224,208),#For static
                     'classifier_column':None,#For static or dynamic
                     'classifier': [True, ('True','False')],
@@ -216,6 +220,7 @@ def create_param_file(filename):
                     'cmap_column':'x',#For dynamic
                     'cmap_max':[100.0,0,1000.0,0.1],#For dynamic
                     'cmap_min':[0.0,0.0,1000.0,0.1],
+                'cmap_name': 'jet',
                     'colour': (0,0,255),#For static
                     'classifier_column':None,#For static or dynamic
                     'classifier': [True, ('True','False')],
@@ -226,6 +231,7 @@ def create_param_file(filename):
                     'cmap_column':'x',#For dynamic                  
                     'cmap_max':[100.0,0,1000.0,0.1],#For dynamic
                     'cmap_min':[0.0,0.0,1000.0,0.1],
+                'cmap_name': 'jet',
                     'classifier_column': None,
                     'classifier': [True, ('True','False')],
                     'colour': (0,255,0),#For static
@@ -236,6 +242,7 @@ def create_param_file(filename):
                     'cmap_column':'voronoi_area',#For dynamic                      'classifier': 1,#For static or dynamic
                     'cmap_max':[100.0,0,1000.0,0.1],#For dynamic
                     'cmap_min':[0.0,0.0,1000.0,0.1],
+                'cmap_name': 'jet',
                     'classifier_column': None,
                     'classifier': [True, ('True','False')],
                     'colour': (0,255,0),#For static
