### 2021-09-23 fe83d41
colormap handles complex values by using np.angle
