### 2021-09-24 #3611 @mikesmithlab
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1016/B978-0-12-391857-4.00009-4 is OK
> - 10.1063/5.0003322 is OK
> - 10.1049/iet-ipr.2018.5684 is OK
> - 10.1016/B978-0-12-391857-4.00009-4 is OK
> - 10.1038/nmeth.2808 is OK
>
> MISSING DOIs
>
> - 10.1016/j.ymeth.2016.09.016 may be a valid DOI for title: TrackMate: An open and extensible platform for single-particle tracking
>
> INVALID DOIs
>
> - None
> ```

Hi I'm not sure what I need to do to fix this. The "Missing DOIs" cites a doi that points to the correct article. However, this is the same DOI which is in my paper.bib file. That DOI also appears correctly in the preprint generated above.
