### 2021-09-24 MikeSmithLabTeam/particletracker#7 @mikesmithlab
There was a small bug in this bit of code which I have fixed in the CurrentDevelopment Branch.
