### 2021-09-24 674f1f3
Double clicking on a particle in the image now highlights the row of the closest particle in the pandas view
