### 2021-09-24 0d2d0f3
Pandas view updates before status bar
