+++ particletracker/gui/main_gui.py
@@ -345,6 +345,13 @@ class MainWindow(QMainWindow):
         output = qimage2ndarray.rgb_view(Qimg)
         print(output[int(y),int(x),:])
 
+        if self.pandas_viewer.isVisible():
+            points = self.pandas_viewer.df[['x', 'y']].values
+            tree = spatial.KDTree(points)
+            dist, idx = tree.query((x, y))
+            self.pandas_viewer.view.selectRow(idx)
+
+
         self.timer = QTimer(self)
         self.timer.timeout.connect(self.reset_statusbar)
         self.timer.start(8000)
@@ -352,11 +359,6 @@ class MainWindow(QMainWindow):
         self.status_bar.showMessage("Coords (x,y): ("+str(x)+','+str(y) +') \t Intensities [r,g,b]:'+str(output[int(y),int(x),:]))    
         self.show()
 
-        if self.pandas_viewer.isVisible():
-            points = self.pandas_viewer.df[['x', 'y']].values
-            tree = spatial.KDTree(points)
-            dist, idx = tree.query((x, y))
-            self.pandas_viewer.view.selectRow(idx)
 
 
 
