### 2021-09-25 #3611 @danielskatz
@mikesmithlab - the problem here is that you have not put some DOIs into the bib file as DOIs, but instead, as URLs.

See https://github.com/MikeSmithLabTeam/particletracker/pull/12 for a fix.  (I'm also removing the URLs that duplicate DOIs, though I think whedon just ignores these in favor of the DOIs anyhow)
