### 2021-09-27 50cecf4
get_outpu removed
