### 2021-09-27 #3611 @mikesmithlab
> @mikesmithlab - the problem here is that you have not put some DOIs into the bib file as DOIs, but instead, as URLs.
>
> See [MikeSmithLabTeam/particletracker#12](https://github.com/MikeSmithLabTeam/particletracker/pull/12) for a fix. (I'm also removing the URLs that duplicate DOIs, though I think whedon just ignores these in favor of the DOIs anyhow)

thanks @danielskatz

@Kevin-Mattheus-Moerman  I think this is all fixed now
