+++ docs/source/index.rst
@@ -16,8 +16,6 @@ Standard use
    rstfiles/batchprocessing
 
 
-
-
 Extending the software
 ----------------------
 .. toctree::
@@ -44,6 +42,19 @@ We aim to test this software against the testdata described in the tutorials how
 If you become aware of an issue please report it https://github.com/MikeSmithLabTeam/particletracker/issues 
 
 
+Citing ParticleTracker
+----------------------
+
+ParticleTracker was created by Mike Smith and James Downs and is offered as open source software which is free for you
+to use. If you use this software for any academic publications please cite this work using the following paper:
+
+
+
+ParticleTracker also relies on two other libraries. Trackpy is used not only for the "trackpy" tracking method but also
+for the linking algorithm. You should therefore also cite this project (). 
+
+OpenCV is used for the contours and hough circles
+tracking methods and the annotation.
 
 Indices and tables
 ==================
