+++ docs/source/rstfiles/tutorials/tutorial7.rst
@@ -1,0 +1,36 @@
+.. _Tutorial7:
+
+Tutorial 7 - Quick Tips and Tricks
+==================================
+
+This tutorial gathers a number of very quick tips and explanations which we think may be useful.
+
+The Mask Tools
+--------------
+
+.. raw:: html
+
+   <iframe width="560" height="315" src="https://www.youtube.com/embed/yVzx76fMqug" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
+    
+
+The Dataviewer
+--------------
+
+.. raw:: html
+
+   <iframe width="560" height="315" src="https://www.youtube.com/embed/R_HBTAx1NeI" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
+    
+Understanding "use_part"
+------------------------
+
+.. raw:: html
+
+   <iframe width="560" height="315" src="https://www.youtube.com/embed/I60HKOYk9zM" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
+
+Working with big files
+----------------------
+
+.. raw:: html
+
+   <iframe width="560" height="315" src="https://www.youtube.com/embed/8yIEVzrLCLQ" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
+
