### 2021-10-01 974877a
Subtract bkg allows subtraction color channel
