### 2021-10-01 c9ec6ba
Added subtract bkg tutorial vid
