+++ docs/source/rstfiles/tutorials/tutorial7.rst
@@ -12,6 +12,13 @@ The Mask Tools
 
    <iframe width="560" height="315" src="https://www.youtube.com/embed/yVzx76fMqug" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
     
+Subtracting a background image
+------------------------------
+
+.. raw:: html
+
+   <iframe width="560" height="315" src="https://www.youtube.com/embed/sWbUAD5FoOE" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write;      encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> 
+
 
 The Dataviewer
 --------------
