### 2021-10-07 3810d55
Added function to extract a dominant frequency from the audio channel of a video and save it in the dataframe
