+++ particletracker/postprocess/postprocessing_methods.py
@@ -4,7 +4,10 @@ import scipy.spatial as sp
 import trackpy as tp
 import cv2
 import os
+import subprocess
 import pandas as pd
+
+from labvision import audio, video
 
 from ..general.parameters import get_method_key, get_param_val
 from ..customexceptions.postprocessor_error import *
@@ -18,15 +21,15 @@ All these methods operate on single frames
 def angle(df, f_index=None, parameters=None, call_num=None):
     '''
     Angle calculates the angle specified by two components.
-    
+
     Notes
     -----
     Usually angle is used following calculating the difference along x and y trajectories.
     It assumes you want to calculate from x_column as dx and y_column as dy
     it uses tan2 so that -dx and +dy give a different result to +dx and -dy
     Angles are output in radians or degrees given by parameters['angle']['units']
-    
-    
+
+
     Parameters
     ----------
 
@@ -38,7 +41,7 @@ def angle(df, f_index=None, parameters=None, call_num=None):
         New column name to store angle df
     units
         'degrees' or 'radians'
-    
+
     Args
     ----
 
@@ -66,16 +69,16 @@ def angle(df, f_index=None, parameters=None, call_num=None):
 
         if output_name not in df.columns:
             df[output_name] = np.nan
-    
+
         df_frame = df.loc[f_index]
-        
+
         if units == 'degrees':
             df_frame[output_name] = np.arctan2(df_frame[columny],df_frame[columnx])*(180/np.pi)
         else:
             df_frame[output_name] = np.arctan2(df_frame[columny],df_frame[columnx])
-    
+
         df.loc[f_index] = df_frame
-    
+
         return df
     except Exception as e:
         raise AngleError(e)
@@ -138,7 +141,7 @@ def classify(df, f_index=None, parameters=None, call_num=None):
 
         df_frame[output_name] = df_frame[column].apply(_classify_fn, lower_threshold_value=lower_threshold_value, upper_threshold_value=upper_threshold_value)
         df.loc[f_index]=df_frame
-   
+
         return df
     except Exception as e:
         raise ClassifyError(e)
@@ -192,7 +195,7 @@ def contour_boxes(df, f_index=None, parameters=None, call_num=None):
     try:
         params = parameters['postprocess']
         method_key = get_method_key('contour_boxes', call_num)
-        
+
         if 'box_cx' not in df.columns:
             df['box_cx'] = np.nan
             df['box_cy'] = np.nan
@@ -201,17 +204,17 @@ def contour_boxes(df, f_index=None, parameters=None, call_num=None):
             df['box_width'] = np.nan
             df['box_area'] = np.nan
             df['box_pts'] = np.nan
-        
+
         df_frame = df.loc[f_index]
         contours = df_frame[['contours']].values
-        
+
         box_cx = []
         box_cy = []
         box_angle = []
         box_length = []
         box_width = []
-        box_area = [] 
-            
+        box_area = []
+
         if np.shape(contours)[0] == 1:
             df_empty = np.isnan(contours[0])
             if np.all(df_empty):
@@ -227,7 +230,7 @@ def contour_boxes(df, f_index=None, parameters=None, call_num=None):
             box_width.append(info_contour[3])
             box_length.append(info_contour[4])
             box_area.append(info_contour[3]*info_contour[4])
-            if index == 0:   
+            if index == 0:
                 box_pts=[info_contour[5]]
             else:
                 box_pts.append(info_contour[5])
@@ -241,7 +244,7 @@ def contour_boxes(df, f_index=None, parameters=None, call_num=None):
         df_frame['box_pts'] = box_pts
 
         df.loc[f_index] = df_frame
-        
+
         return df
[…]
