### 2021-10-21 8b9c3e3
Corrected bug when finding only 0 or 1 particles
