+++ particletracker/postprocess/postprocessing_methods.py
@@ -67,14 +67,14 @@ def angle(df, f_index=None, parameters=None, call_num=None):
         if output_name not in df.columns:
             df[output_name] = np.nan
     
-        df_frame = df.loc[f_index]
+        df_frame = df.loc[[f_index]]
         
         if units == 'degrees':
             df_frame[output_name] = np.arctan2(df_frame[columny],df_frame[columnx])*(180/np.pi)
         else:
             df_frame[output_name] = np.arctan2(df_frame[columny],df_frame[columnx])
     
-        df.loc[f_index] = df_frame
+        df.loc[[f_index]] = df_frame
     
         return df
     except Exception as e:
@@ -134,10 +134,10 @@ def classify(df, f_index=None, parameters=None, call_num=None):
 
         if output_name not in df.columns:
             df[output_name] = np.nan
-        df_frame = df.loc[f_index]
-
+        df_frame = df.loc[[f_index]]
+      
         df_frame[output_name] = df_frame[column].apply(_classify_fn, lower_threshold_value=lower_threshold_value, upper_threshold_value=upper_threshold_value)
-        df.loc[f_index]=df_frame
+        df.loc[[f_index]]=df_frame
    
         return df
     except Exception as e:
@@ -202,7 +202,7 @@ def contour_boxes(df, f_index=None, parameters=None, call_num=None):
             df['box_area'] = np.nan
             df['box_pts'] = np.nan
         
-        df_frame = df.loc[f_index]
+        df_frame = df.loc[[f_index]]
         contours = df_frame[['contours']].values
         
         box_cx = []
@@ -240,7 +240,7 @@ def contour_boxes(df, f_index=None, parameters=None, call_num=None):
         df_frame['box_area'] = box_area
         df_frame['box_pts'] = box_pts
 
-        df.loc[f_index] = df_frame
+        df.loc[[f_index]] = df_frame
         
         return df
     except Exception as e:
@@ -301,10 +301,10 @@ def logic_AND(df, f_index=None, parameters=None, call_num=None):
         output_name = params[method_key]['output_name']
         if output_name not in df.columns:
             df[output_name] = np.nan
-        df_frame = df.loc[f_index]
+        df_frame = df.loc[[f_index]]
         
         df_frame[output_name] = df_frame[column1] * df_frame[column2]
-        df.loc[f_index] = df_frame
+        df.loc[[f_index]] = df_frame
         
         return df
     except Exception as e:
@@ -352,10 +352,10 @@ def logic_NOT(df, f_index=None, parameters=None, call_num=None):
         output_name = params[method_key]['output_name']
         if output_name not in df.columns:
             df[output_name] = np.nan
-        df_frame = df.loc[f_index]
+        df_frame = df.loc[[f_index]]
 
         df_frame[output_name] = ~df_frame[column]
-        df.loc[f_index] = df_frame
+        df.loc[[f_index]] = df_frame
         return df
     except Exception as e:
         raise LogicNotError(e)
@@ -404,10 +404,10 @@ def logic_OR(df, f_index=None, parameters=None, call_num=None):
 
         if output_name not in df.columns:
             df[output_name] = np.nan
-        df_frame = df.loc[f_index]
+        df_frame = df.loc[[f_index]]
 
         df_frame[output_name] = df_frame[column1] + df_frame[column2]
-        df.loc[f_index] = df_frame
+        df.loc[[f_index]] = df_frame
         return df
     except  Exception as e:
         raise LogicOrError(e)
@@ -453,10 +453,10 @@ def magnitude(df, f_index=None, parameters=None, call_num=None):
         
         if output_name not in df.columns:
             df[output_name] = np.nan
[…]
