+++ README.md
@@ -30,9 +30,7 @@ HTML:
   <img src="https://joss.theoj.org/papers/10.21105/joss.03611/status.svg" alt="DOI badge" >
 </a>
 
-reStructuredText:
-.. image:: https://joss.theoj.org/papers/10.21105/joss.03611/status.svg
-   :target: https://doi.org/10.21105/joss.03611
+
    
 ## Contributions
 Details about contributing to the ParticleTracker project can be found [here](https://github.com/MikeSmithLabTeam/particletracker/blob/master/CONTRIBUTING.md)
