+++ README.md
@@ -22,13 +22,12 @@ Add the following argument to setup.py setuptools.setup()
     dependency_links=['https://github.com/MikeSmithLabTeam/particletracker/tarball/repo/master#egg=package-1.0'],
     
 ## Citation
+To cite this project in your publications please cite the following paper:
+"ParticleTracker: a gui based particle tracking software"
+M.I. Smith, J.G. Downs, J. Open Source Software 6, 3611 (2021)
+
 Markdown:
 [![DOI](https://joss.theoj.org/papers/10.21105/joss.03611/status.svg)](https://doi.org/10.21105/joss.03611)
-
-HTML:
-<a style="border-width:0" href="https://doi.org/10.21105/joss.03611">
-  <img src="https://joss.theoj.org/papers/10.21105/joss.03611/status.svg" alt="DOI badge" >
-</a>
 
 
    
