+++ docs/source/rstfiles/introduction.rst
@@ -37,10 +37,6 @@ to use. If you use this software for any academic publications please cite this 
 
 "ParticleTracker: a gui based particle tracking software" Journal of Open Source Software 6, 3611 (2021), M.I. Smith, J.G. Downs."
 
-Markdown:
-[![DOI](https://joss.theoj.org/papers/10.21105/joss.03611/status.svg)](https://doi.org/10.21105/joss.03611)
-
-
 
 ParticleTracker also relies on two other libraries. Trackpy is used not only for the "trackpy" tracking method but also
 for the linking algorithm. You should therefore also cite this project (https://zenodo.org/record/4682814#.YVcuc9rMLIU). 
