+++ paper.bib
@@ -4,7 +4,7 @@
   year = {2016},
   publisher = {Elsevier},
   journal = {Methods},
-  url = {https://doi.org/10.1016/j.ymeth.2016.09.016}
+  doi = {10.1016/j.ymeth.2016.09.016}
 }
 
 
@@ -14,7 +14,7 @@
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://zenodo.org/record/3492186#.YUMkp7hKjIV}
+  doi = {10.5281/zenodo.3492186}
 }
 
 @article{opencv_library,
@@ -36,7 +36,6 @@
     year = 2012,
     volume = 504,
     doi = {10.1016/B978-0-12-391857-4.00009-4},
-    url = {https://www.sciencedirect.com/science/article/pii/B9780123918574000094?via%3Dihub},
 }
 
 @article{Nanoparticles,
@@ -47,7 +46,6 @@
     year = 2020,
     volume = 127,
     doi = {10.1063/5.0003322},
-    url = {https://aip.scitation.org/doi/10.1063/5.0003322},
 }
 
 @article{Foams,
@@ -58,7 +56,6 @@
     year = 2019,
     volume = 13,
     doi = {10.1049/iet-ipr.2018.5684},
-    url = {https://ietresearch.onlinelibrary.wiley.com/doi/10.1049/iet-ipr.2018.5684},
 }
 
 
@@ -71,7 +68,6 @@
     year = 2012,
     volume = 504,
     doi = {10.1016/B978-0-12-391857-4.00009-4},
-    url = {https://www.sciencedirect.com/science/article/pii/B9780123918574000094?via%3Dihub},
 }
 
 @misc{RyTrack,
@@ -88,8 +84,5 @@
     year = 2014,
     volume = 11,
     doi = {10.1038/nmeth.2808},
-    url = {https://www.nature.com/articles/nmeth.2808},
 }
-
-
 
