+++ docs/source/rstfiles/tutorials.rst
@@ -1,5 +1,5 @@
 Tutorials
-========
+=========
 
 In these examples we illustrate how to build a number of different tracking projects.
 
