### 2021-10-27 16e1a11
try removing autodoc for postprocessing
