+++ docs/source/rstfiles/reference/postprocess_ref.rst
@@ -1,5 +1,3 @@
 Postprocessing
 ==============
 
-.. automodule:: particletracker.postprocess.postprocessing_methods
-    :members:
