+++ docs/requirements.txt
@@ -1,2 +1,2 @@
-sphinx>=1.7.5
+sphinx>=4.2.0
 PyQt5
