+++ docs/source/rstfiles/reference/postprocess_ref.rst
@@ -1,3 +1,5 @@
 Postprocessing
 ==============
 
+.. automodule:: particletracker.postprocess.postprocessing_methods
+    :members:
