### 2021-11-10 a6ca708
corrected naming of batchprocess in docs
