### 2021-11-22 23ec867
Update main_gui.py
setup_pandas_viewer calling self.update_pandas_view() can cause crash when loading a settings file. It is also unnecessary as the update_viewer also calls update_pandas_viewer.
