### 2021-12-07 7c16192
added postprocessing method for Re and Imag parts
