### 2021-12-07 4e129b8
added method to calc absolute value of column
