### 2021-12-07 ec4854c
added absolute postprocessing method
