+++ particletracker/general/param_file_creator.py
@@ -138,8 +138,11 @@ def create_param_file(filename):
         'hexatic_order':{
             'threshold': [10, 1, 100, 1]
         },
-        'Re_Im_Components':{
+        're_im_components':{
             'column_name': 'hexatic_order'
+        },
+        'absolute':{
+            'column_name':'x'
         }
 
         }
