### 2021-12-08 3dfcd75
added complex magnitude to real_imag
