### 2021-12-08 35d4273
making naming conventions of errors consistent
