### 2021-12-08 5d04b49
forced number_of_neighbours to be real
