### 2021-12-16 3859679
remove_masked moved postprocessor to hough track
