+++ particletracker/postprocess/postprocessing_methods.py
@@ -657,97 +657,6 @@ def _get_class_subset(df, f, parameters, method=None):
     return subset_df
 
 
-def remove_masked(df, f_index=None, parameters=None, call_num=None):
-    '''
-    Remove masked objects
-    
-    Notes
-    -----
-    The Hough circles tracking method can find circles with centres 
-    outside the masked area. This method enables you to remove those 
-    points from the data.
-    
-    Parameters
-    ----------
-    column_name
-        Input column names
-    output_name
-        Output column name
-    fps
-        numerical value indicating the number of frames per second
-    span
-        number of frames over which to calculate rolling difference
-    
-    
-    
-    Args
-    ----
-
-    df
-        The dataframe in which all data is stored
-    f_index
-        Integer specifying the frame for which calculations need to be made.
-    parameters
-        Nested dictionary like object (same as .param files or output from general.param_file_creator.py)
-    call_num
-        Usually None but if multiple calls are made modifies method name with get_method_key
-
-    Returns
-    -------
-        updated dataframe
-
-
-    '''
-    try:
-        params = parameters['postprocess']
-        method_key = get_method_key('remove_masked', call_num)
-
-        mask_method_list = list(parameters['crop']['crop_method'])
-        if 'crop_box' in mask_method_list: mask_method_list.remove('crop_box')
-
-        contour_list = []
-        contour_list.append(_contour_from_mask(parameters['crop'][mask_method_list[0]],mask_method_list[0]))
-
-        df_frame = df.loc[f_index]
-        points = df_frame[['x','y']].apply(tuple, axis=1).values
-        mask = np.array([_point_inside_mask(point, contour_list) for point in points])
-        for column in df_frame.columns:
-            df_frame[column] = df_frame[column].where(mask)
-        df_frame = df.loc[[f_index]]
-        df.dropna(how='all',inplace=True)
-        return df
-    except Exception as e:
-        raise RemoveMaskedError(e)
-
-def _contour_from_mask(mask_pts, mask_type):
-    if mask_type == 'mask_rectangle':
-        x1 = mask_pts[0][0]
-        x2 = mask_pts[1][0]
-        y1 = mask_pts[0][1]
-        y2 = mask_pts[1][1]
-
-        contour = [np.array([[x1,y1],[x1,y2],[x2,y2],[x2, y1]])]
-    elif mask_type == 'mask_ellipse':
-        pass
-    elif mask_type == 'mask_circle':
-        pass
-    elif mask_type == 'mask_polygon':
-        pass
-    else:
-        print('Error unrecognised mask type')
-        raise Exception
-    return contour
-
-def _point_inside_mask(point, mask_contour_list):
-    inside = False
-    for contour in mask_contour_list:
-
-        result = cv2.pointPolygonTest(contour[0], point, False)
-        if result != -1:
-            inside = True
-    return inside
-
-
 def hexatic_order(df, f_index=None, parameters=None, call_num=None):
     """
     Calculates the hexatic order parameter of each particle. Neighbours are 
