+++ particletracker/customexceptions/postprocessor_error.py
@@ -120,12 +120,6 @@ class AbsoluteError(PostprocessorError):
         self.error_msg = 'absolute error'
         self.e = e
 
-
-class RemoveMaskedError(PostprocessorError):
-    def __init__(self, e):
-        super().__init__(e)
-        self.error_msg = "remove masked objects error"
-
 class AudioFrequencyError(PostprocessorError):
     def __init__(self, e):
         super().__init__(e)
