### 2021-12-17 9a5d2a3
duplicate methods need deepcopy upon creation
