### 2022-01-07 452a6e7
Added fix for drag_drop_list
